requirements.txt
py-bcrypt
python-decouple
simple-crypt
